Removed rthe rest of the vitess specific functionality, as well as the analyzer (which we weren't using)
